# 增加元素添加页面 # 增加ui用例 # 增加ui调试 # 增加ui用例报告 # 增加异步任务执行记录

+++ frontend/src/views/ui/page/uiElement.vue
@@ -1,104 +1,25 @@
 <template>
   <el-card>
-    <template #header>
-      <div style="display: flex; justify-content: space-between">
-        <strong>元素管理</strong>
-        <div>
-          <el-button type="primary" @click="addElement" :disabled="!props.pageId">添加元素</el-button>
-          <el-button type="success" @click="getElementList" :disabled="!props.pageId">
-            <el-icon>
-              <Refresh></Refresh>
-            </el-icon>
-            刷新
-          </el-button>
-        </div>
-      </div>
-    </template>
-
-    <el-table
+    <z-table
+        :columns="state.columns"
         :data="state.elementList"
-        v-loading="state.loading"
+        ref="tableRef"
+        :border="false"
+        v-model:page-size="state.listQuery.pageSize"
+        v-model:page="state.listQuery.page"
+        :total="state.total"
+        @pagination-change="getElementList"
     >
-      <template v-for="(col, index) in state.columns" :key="col.key">
-
-        <el-table-column
-            v-if="col.key === 'location_type'"
-            :show-overflow-tooltip="col.show"
-            :label="col.label"
-            :sortable="col.sortable"
-            :prop="col.key"
-            :fixed="col.fixed"
-            :align="col.align"
-            :width="col.width">
-          <template #default="{row}">
-            <el-select v-model="row.location_type"
-                       placeholder="请选择"
-                       style="width:100%"
-                       v-if="row._edit">
-              <el-option
-                  v-for="type in state.locationTypes"
-                  :key="type.value"
-                  :label="type.label"
-                  :value="type.value">
-              </el-option>
-            </el-select>
-            <span v-else>{{ row[col.key] }}</span>
-          </template>
-        </el-table-column>
-
-        <el-table-column
-            v-else
-            :show-overflow-tooltip="col.show"
-            :label="col.label"
-            :sortable="col.sortable"
-            :prop="col.key"
-            :fixed="col.fixed"
-            :align="col.align"
-            :width="col.width">
-          <template #default="{row}">
-            <el-input v-if="row._edit" v-model="row[col.key]"></el-input>
-            <span v-else>{{ row[col.key] }}</span>
-          </template>
-        </el-table-column>
-
-      </template>
-
-      <el-table-column
-          show-overflow-tooltip="true"
-          label="操作"
-          align="center"
-          width="120">
-        <template #default="{row, $index}">
-          <div v-show="!row._edit">
-            <el-button type="primary" @click="editElement(row)">编辑</el-button>
-            <el-button type="danger" @click="deletedElement(row)">删除</el-button>
-          </div>
-          <div v-show="row._edit">
-            <el-button type="warning" @click="saveElement(row)">保存</el-button>
-            <el-button type="info" @click="closeEditElement(row, $index)">取消</el-button>
-          </div>
-        </template>
-      </el-table-column>
-    </el-table>
-
-    <div class="mt20">
-      <el-pagination
-          :total="state.total"
-          :page-size="state.listQuery.pageSize"
-          layout="total, sizes, prev, pager, next, jumper"
-          v-model:current-page="state.listQuery.page"
-          @size-change="getElementList"
-      />
-    </div>
-
+    </z-table>
   </el-card>
 </template>
 
 <script setup lang="ts" name="UiElement">
-import {ElButton, ElMessage, ElMessageBox} from "element-plus";
-import {h, onMounted, reactive, watch} from "vue";
+import {ElButton, ElInput, ElMessage, ElMessageBox, ElOption, ElSelect} from "element-plus";
+import {h, reactive, watch} from "vue";
 import {useUiElementApi} from "/@/api/useUiApi/uiElement";
-import {Refresh} from "@element-plus/icons-vue";
+import {RefreshLeft, Edit, CirclePlus} from "@element-plus/icons-vue"
+
 
 const props = defineProps({
   pageId: {
@@ -109,19 +30,88 @@ const props = defineProps({
 
 const state = reactive({
   columns: [
+    {label: '序号', columnType: 'index', width: 'auto', show: true},
     {
-      key: 'name', label: '元素名称', width: '', show: true,
-      render: (row: any) => h(ElButton, {
-        link: true,
-        type: "primary",
-        onClick: () => {
-          onOpenSaveOrUpdate("update", row)
-        }
-      }, () => row.name)
+      key: 'name', label: '元素名称', width: '', align: 'center', show: true,
+      render: ({row}: any) => h(ElInput, {
+        modelValue: row.name,
+        placeholder: '请输入元素名称',
+        clearable: true,
+      }),
     },
-    {key: 'location_type', label: '定位类型', width: '', align: 'center', show: true},
-    {key: 'location_value', label: '定位值', width: '', align: 'center', show: true},
-    {key: 'remarks', label: '备注', width: '', align: 'center', show: true},
+    {
+      key: 'location_method', label: '定位方式', width: '', align: 'center', show: true,
+      render: ({row}: any) => h(ElSelect, {
+          style: 'width: 100%',
+          modelValue: row.location_method,
+          placeholder: '请选择定位方式',
+          clearable: true,
+        }, () => state.locationTypes.map((type: any) => h(ElOption, {
+          key: type.value,
+          label: type.label,
+          value: type.value
+        }))
+      ),
+      renderHeader: () => h('div', {}, [
+        h('span', {
+          style: 'margin-right: 10px'
+        }, "定位方式"),
+        h(ElButton, {
+          type: "primary",
+          icon: Edit,
+          round: true,
+        }, () => "编辑"),
+        h(ElButton, {
+          type: "success",
+          icon: RefreshLeft,
+          round: true,
+        }, () => "刷新")
+      ])
+    },
+    {
+      key: 'location_value', label: '定位值', width: '', align: 'center', show: true,
+      render: ({row}: any) => h(ElInput, {
+        modelValue: row.location_value,
+        placeholder: '请输入定位值',
+        clearable: true,
+      },)
+    },
+    {
+      key: 'remarks', label: '备注', width: '', align: 'center', show: true,
+      render: ({row}: any) => h(ElInput, {
+        modelValue: row.remarks,
+        placeholder: '请输入备注',
+        clearable: true,
+      },)
+    },
+    {
+      label: '操作', fixed: 'right', width: '140', align: 'center',
+      render: ({row, $index}: any) => h("div", null, [
+        h(ElButton, {
+          type: "primary",
+          onClick: () => {
+            saveElement(row)
+          }
+        }, () => '保存'),
+
+        h(ElButton, {
+          type: "danger",
+          onClick: () => {
+            deletedElement(row, $index)
+          }
+        }, () => '删除')
+      ]),
+      renderHeader: () => h('div', {}, [
+        h(ElButton, {
+          type: "success",
+          icon: CirclePlus,
+          round: true,
+          onClick: () => {
+            addElement()
+          }
+        }, () => "添加元素")
+      ])
+    },
   ],
   elementList: [],
   total: 0,
@@ -145,46 +135,28 @@ const state = reactive({
     {label: 'title', value: 'title'},
   ],
 
-//   tmpElRows
-  tmpElRows: {},
-
 });
 
 const getElementList = () => {
   state.loading = true
   useUiElementApi().getList(state.listQuery)
-      .then((res: any) => {
-        state.elementList = res.data.rows;
-        state.total = res.data.rowTotal;
-      })
-      .finally(() => {
-        state.loading = false
-      })
+    .then((res: any) => {
+      state.elementList = res.data.rows;
+      state.total = res.data.rowTotal;
+    })
+    .finally(() => {
+      state.loading = false
+    })
 };
 
-// 编辑元素
-const editElement = (row: any) => {
-  state.tmpElRows[row.id] = {...row}
-  row._edit = true
-}
-
-// 关闭编辑状态
-const closeEditElement = (row: any, index: number) => {
-  if (row.id) {
-    row._edit = false
-  } else {
-    state.elementList.splice(index, 1)
-  }
-}
 
 const addElement = () => {
   state.elementList.unshift({
     page_id: props.pageId,
     name: '',
-    location_type: '',
+    location_method: '',
     location_value: '',
     remarks: '',
-    _edit: true,
   })
 
 }
@@ -193,17 +165,17 @@ const addElement = () => {
 const saveElement = (row: any) => {
   console.log(state, "state")
   // 判断有 row.page_name 则保存，不然就是url为空
-  if (row.location_type && row.name) {
+  if (row.location_method && row.name) {
     useUiElementApi().saveOrUpdate(row)
-        .then(res => {
-          row._edit = false
-          ElMessage.success('保存成功')
-          // 保存成功后不刷新页面，直接塞入id， 不然后面点删除数据id为空
-          if (!row.id) {
-            state.total++
-          }
-          row.id = res.data.id
-        })
+      .then(res => {
+        row._edit = false
+        ElMessage.success('保存成功')
+        // 保存成功后不刷新页面，直接塞入id， 不然后面点删除数据id为空
+        if (!row.id) {
+          state.total++
+        }
+        row.id = res.data.id
+      })
   } else {
     ElMessage.warning('元素，或者定位值不能为空!')
   }
@@ -211,32 +183,36 @@ const saveElement = (row: any) => {
 
 
 // 删除
-const deletedElement = (row: any) => {
+const deletedElement = (row: any, index: number) => {
   ElMessageBox.confirm('是否删除该条数据, 是否继续?', '提示', {
     confirmButtonText: '确定',
     cancelButtonText: '取消',
     type: 'warning'
   }).then(() => {
+    if (!row.id) {
+      state.elementList.splice(index, 1)
+      return
+    }
     useUiElementApi().deleted({id: row.id})
-        .then((res: any) => {
-          ElMessage.success('删除成功')
-          getElementList()
-        })
+      .then((res: any) => {
+        ElMessage.success('删除成功')
+        getElementList()
+      })
   })
 }
 
 watch(
-    () => props.pageId,
-    (page_id) => {
-      state.elementList = []
-      if (page_id) {
-        state.listQuery.page_id = props.pageId
-        getElementList()
-      }
-    },
-    {
-      deep: true,
+  () => props.pageId,
+  (page_id) => {
+    state.elementList = []
+    if (page_id) {
+      state.listQuery.page_id = props.pageId
+      getElementList()
     }
+  },
+  {
+    deep: true,
+  }
 );
 
 </script>
@@ -259,6 +235,10 @@ watch(
   :deep(.input-with-select .el-input-group__prepend) {
     background-color: var(--el-fill-color-blank);
   }
+}
+
+:deep(.el-table .el-table__cell) {
+  padding: 8px 0;
 }
 
 </style>
